pyproject.toml
[project]
name = "experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tdqm>=0.0.1",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
